remove the public account and apikey support
