+++ chatbees/client/admin_management.py
@@ -7,22 +7,20 @@ from chatbees.server_models.ingestion_api import (
 )
 from chatbees.utils.config import Config
 
-__all__ = ["init", "create_api_key", "list_connectors"]
+__all__ = ["init", "list_connectors"]
 
 
 def init(
-    api_key: str = None,
-    account_id: str = Config.PUBLIC_ACCOUNT_ID,
+    api_key: str,
+    account_id: str,
     namespace: str = Config.PUBLIC_NAMESPACE
 ):
     """
     Initialize the ChatBees client.
 
     Args:
-        api_key (str, optional): The API key to authenticate requests.
-                                 Defaults to None. You can still access
-                                 public collections without an API key.
-        account_id (str, optional): The account ID. Defaults to "public".
+        api_key (str): The API key to authenticate requests.
+        account_id (str): The account ID.
         namespace (str, optional): The namespace to use.
     Raises:
         ValueError: If the provided config is invalid
@@ -41,25 +39,3 @@ def list_connectors() -> List[ConnectorReference]:
         data=req.model_dump_json(),
     )
     return ListConnectorsResponse.model_validate(resp.json()).connectors
-
-
-def create_api_key() -> str:
-    """
-    Create a new APIKey in the public account.
-
-    Returns:
-        The created API key
-
-    """
-    if Config.account_id != Config.PUBLIC_ACCOUNT_ID:
-        raise ValueError(
-            "Please log in and create API key on UI for the private account")
-
-    url = f'{Config.get_base_url()}/apikey/create'
-    req = CreateApiKeyRequest()
-    resp = Config.post(
-        url=url,
-        data=req.model_dump_json(),
-        enforce_api_key=False,
-    )
-    return CreateApiKeyResponse.model_validate(resp.json()).api_key
