+++ chatbees/tests/regression_test.py
@@ -6,9 +6,11 @@ import uuid
 from typing import List, Any
 
 import chatbees as cb
+from chatbees.server_models.doc_api import AnswerReference
 
 TEST_AID = os.environ.get('ENV_TEST_AID')
 TEST_APIKEY = os.environ.get('ENV_TEST_APIKEY')
+TEST_PUBLIC_AID = os.environ.get('ENV_TEST_PUBLIC_AID')
 TEST_PUBLIC_APIKEY = os.environ.get('ENV_TEST_PUBLIC_APIKEY')
 
 TEST_CONFLUENCE_USER = os.environ.get('TEST_CONFLUENCE_USER')
@@ -56,7 +58,7 @@ class RegressionTest(unittest.TestCase):
         logging.info(f"refs={doc_names}")
 
     def test_regress_public(self):
-        cb.init(TEST_PUBLIC_APIKEY)
+        cb.init(TEST_PUBLIC_APIKEY, TEST_PUBLIC_AID)
         # read-only
         clname1 = 'llm_research_test'
         clname2 = 'regress_pngame_docs'
@@ -65,6 +67,74 @@ class RegressionTest(unittest.TestCase):
         clname1 = 'llm_research_test_write'
         clname2 = 'regress_pngame_docs_write'
         self._test_doc_and_web(clname1, clname2, write=True)
+
+    def test_doc_apis(self):
+        cb.init(api_key=self.apikey, account_id=self.aid)
+        clname = 'test_doc_apis'
+        col = cb.Collection(name=clname)
+        cb.create_collection(col)
+
+        files = [
+            f'{os.path.dirname(os.path.abspath(__file__))}/data/text_file.txt',
+            f'{os.path.dirname(os.path.abspath(__file__))}/data/española.txt',
+            f'{os.path.dirname(os.path.abspath(__file__))}/data/française.txt',
+            f'{os.path.dirname(os.path.abspath(__file__))}/data/中文.txt',
+        ]
+        doc_names = {'text_file.txt', 'española.txt', 'française.txt', '中文.txt'}
+
+        try:
+            # add and summarize
+            for file in files:
+                col.upload_document(file)
+                fname = os.path.basename(file)
+                col.summarize_document(fname)
+
+            # list
+            print("list_documents")
+            list_doc_names = col.list_documents()
+            assert doc_names == set(list_doc_names)
+
+            # ask
+            print("ask")
+            resp = col.ask('question?')
+            assert len(resp.refs) > 0
+
+            # delete, then list and ask again
+            col.delete_document('española.txt')
+            doc_names = {'text_file.txt', 'française.txt', '中文.txt'}
+
+            list_doc_names = col.list_documents()
+            assert doc_names == set(list_doc_names)
+
+            resp = col.ask('question?')
+            assert len(resp.refs) > 0
+
+            # chat
+            chat1 = col.chat()
+            chat2 = col.chat(doc_name="text_file.txt")
+
+            chat1.ask("q1")
+            chat1.ask("q2")
+            chat1.ask("q3")
+
+            resp = chat2.ask("q1")
+            self.assertRefsAreFromDoc(resp.refs, "text_file.txt")
+            resp = chat2.ask("q2")
+            self.assertRefsAreFromDoc(resp.refs, "text_file.txt")
+            resp = chat2.ask("q3")
+            self.assertRefsAreFromDoc(resp.refs, "text_file.txt")
+
+            # ensure we can configure chat attrs
+            col.configure_chat('a pirate from 1600s', 'the word snowday and nothing else')
+            resp = col.ask('what is the color of my hair?')
+            print("persona answer", resp.answer)
+        finally:
+            cb.delete_collection(col.name)
+
+    def assertRefsAreFromDoc(self, refs: List[AnswerReference], doc: str):
+        assert len(refs) > 0
+        for ref in refs:
+            assert ref.doc_name == doc
 
     def test_regress_acc(self):
         cb.init(api_key=self.apikey, account_id=self.aid)
@@ -85,7 +155,6 @@ class RegressionTest(unittest.TestCase):
 
         self._test_confluence_user()
         self._test_connector_ingests()
-
 
     def _test_doc_and_web(self, clname1: str, clname2: str, write: bool = True):
         cols = cb.list_collections()
@@ -216,3 +285,44 @@ class RegressionTest(unittest.TestCase):
         resp = col.ask(q)
         logging.info(f"Question: {q}")
         logging.info(f"Answer: {resp.answer}\n")
+
+
+    """
+    def test_collection_apis(self):
+        # Clear API key from config
+        apikey1 = self.apikey1
+        apikey2 = self.apikey2
+
+        # Create private collections, one for each created key
+        cb.init(api_key=apikey1)
+        private_col_key1 = self.create_collection()
+
+        cb.init(api_key=apikey2)
+        private_col_key2 = self.create_collection()
+
+        try:
+            # List collections using API key1
+            cb.init(api_key=apikey1)
+            collections_visible_to_key1 = set(cb.list_collections())
+            assert private_col_key2.name not in collections_visible_to_key1
+            assert private_col_key1.name in collections_visible_to_key1
+            # Key1 is not authorized to delete a collection created by key2
+            self.assertRaises(cb.UnAuthorized, cb.delete_collection, private_col_key2.name)
+
+            # List collections using API key2
+            cb.init(api_key=apikey2)
+            collections_visible_to_key2 = set(cb.list_collections())
+            assert private_col_key1.name not in collections_visible_to_key2
+            assert private_col_key2.name in collections_visible_to_key2
+            # Key2 is not authorized to delete a collection created by key1
+            self.assertRaises(cb.UnAuthorized, cb.delete_collection, private_col_key1.name)
+
+        finally:
+            # key1 is authorized to delete its own collections as well as public
+            # collections
+            cb.init(api_key=apikey1)
+            cb.delete_collection(private_col_key1.name)
+
+            cb.init(api_key=apikey2)
+            cb.delete_collection(private_col_key2.name)
+    """
